cleanup-service/requirements.txt
requests
python-dotenv
redis
rq
sqlalchemy
mysql-connector-python
tenacity

